signin/requirements.txt
# Function dependencies, for example:
# package>=version
pymongo
Flask
google-auth
google-auth-oauthlib
google-api-python-client
python-dotenv
